requirements.txt
djangorestframework
Django>=4.0
djangorestframework-simplejwt
django-filter
